fix template types of Threshold::thresh and Threshold::type. remove regacy include in LabeledObject.cpp.

+++ include/Threshold.h
@@ -6,9 +6,9 @@ namespace cvImagePipeline {
 		{
 		public:
 			DECLARE_CVFILTER;
-			TParam<int> thresh;
+			TParam<double> thresh;
 			TParam<double> maxval;
-			TParam<double> type;
+			TParam<int> type;
 		public:
 			Threshold();
 			virtual ~Threshold();
